Turn small series optimizations on by default
The ability to fully fit the data payload inside of a REBSER has moved
in phases from only being legal for small fixed size arrays, to being
legal for small fixed size arbitrary series.  This pushes to the next
step which is that the optimization is applied whether it is requested
or not...so small series or arrays of length 0 or 1 will be packed
into the REBSER automatically, and able to resize if needed.

Because this required changing quite a bit of code, the opportunity
was taken to go ahead and apply the growing trend that in functions
which talk about one series only to just call it `s` in order to make
it more readable.  Since basically every usage of a series variable
requires SER_HEAD(...) etc, the readability of the relevant parts of
the code goes up to see that as SER_HEAD(s) instead of SER_HEAD(series)
in repeated expressions.

Although the behavior is automatic, this still keeps the singular
array allocator, since there is a small optimization in the sense that
it doesn't need to go through a 0-length step.  However, this applies
the fixed-size bit...so it should only be used if it never plans to
grow or shrink to 0.