Make CHAINed functions hold argument array live
In order to appease the error machinery, the dropping of arguments to
a chained input function was held off until after the chain was done.
The shift of the eval_type away from holding those arguments live
during the chain was not taken out with that.

It remains a question as to what the best answer is, but if the
arguments are live the evaluation type has to stay as a function--or
there will be unwanted GCs, causing a crash when the drop happens.