Break include files up into smaller units
The requirements for inlining have caused more structure to emerge from
what was once a fairly random ordering of macros in R3-Alpha (since the
order did not matter, as they were not checked until they were inlined).

This takes another step toward organization, breaking some of the files
out based on their functional grouping.  This means that definitions
for the array REBSER subclass, as well as for the value type using
that subclass, can be grouped together.