Add LOGIC! to legal arguments to ASSERT
While allowing any argument type to ASSERT would be asking for trouble
(e.g. whether to run a BLOCK! that came back from an evaluation or
regard it as TRUE?), being able to pass a LOGIC! without a block is
reasonable.  One can thus write:

    assert any? [a b c]

Instead of:

    assert [any [a b c]]

It will run a bit faster and is less visually cluttering.