Patches for syncing up Ren Garden
Ren-Cpp compiles with stricter settings than Ren-C, and notices a few
more error categories.  So when functions get moved to being inlined,
that means picking up on those errors in Ren-Cpp.

Also observed that END values in cells handed out via API (in pairings)
need to be legal.  Otherwise it's not possible to DO into an API cell.