64-bit alignment audit, improve comments in memory code
While doing some comment cleanup in the %m-pools.c file, it became
apparent that the 64-bit alignment guarantee for REBVAL on 32-bit
platforms had not been maintained by some of the tricks that were
putting cells into REBSER payloads, or in the chunk stack.

In order to get on the right side of that, this moves the `link` field
in the REBSER to before the content...so that two platform pointers
come after the REBNOD allocation point (which is 64-bit aligned).
Then the chunk stack offset and size were merged into one field, since
the offset doesn't have to be that large to begin with...this meant
there were only two pointers from the start of the chunk to the payload
and hence 32 + 32 = 64 if a pointer is 32 bits.

While tending these issues, this also centralizes notes about the
header aliasing and why that happens...by using a common inlined routine
to do the initialization through a Reb_Header pointer alias.