Inline data stack macros, update comments, allow voids
There were some outdated remarks in the comments about the data stack
and chunk stack, as well as unsafe macros that repeated their
arguments (and a comment saying to fix that once inline functions had
been sorted out).  This does that.

Also, there was a disallowance of pushing voids.  This is applicable
to some clients but not all, and wouldn't be checked by a DS_PUSH_TRASH
followed by a SET_VOID().  So checking it in the push itself doesn't
have much teeth, and a garbage collection will catch it and indicate
the exact location and tick the void came from.  Deemed not worth it
and removed for simplification.