Get experimental debug interface running again
The debugger primitives had a failed shakeup in being written as
mostly Rebol, which was backed out.  This means the C-heavy debug
routines for showing backtraces and selecting frames is still C.

That code had been allowed to atrophy in the meantime.  This gets it
back into shape so the old demos could be demonstrated at Atronix.

(Even though the interface for these routines is not set in stone, it
would probably be worthwhile to write some tests for what's there,
just to notice when regressions occur.)