Get rid of NULL checking in Action_Dispatcher
The "action" dispatcher sniffs the type of the first argument to the
action, and uses that type to index into a table of per-type handlers
that are like dispatchers, but take the canon value of a symbol as an
additional argument to decode what to do.  (The T_XXX functions).

The only reason it was checking for NULL was because to deliver an
error message if it dispactched to a HANDLE!.  By simply making a
dispatcher for the T_Handle that gives an error, this removes the need
for the NULL check.

Includes a patch in the tests for a QUIT/RETURN that somehow stuck
around after the /RETURN compatibility for QUIT/WITH was moved to
the r3-legacy.