Don't clear line flag on words when rebinding
When a word was being bound relatively, the header was being reset.
This was in order to clear out any old binding flags before making any
new ones.  However, this cleared the line flag--and could create other
problems too.

This changes the code to use UNBIND_WORD, which is tolerant even if
the word was not previously bound.