Eliminate compiling warning
Because GC_Ballast is I64, passing &GC_Ballast to REB_I32_ADD_OF will
cause an imcompatible pointer warning.

Because size is 32-bit integer, adding it to a 64-bit integer will not
overflow the 64-bit integer, as long as the 64-bit integer is less than
MAX_I32. There is no need to check for overflow here.