Replace HAS-TYPE? specializations with TYPECHECKER
Checking against types or typesets was previously done by way of
specializing an arity-2 HAS-TYPE? function, which took both a data
type and a value to check.  So during bootstrap, a loop would make
a specialization for each type, something like:

     string?: specialize :has-type? [type: string!]

However, there is a small cost to the specializations compared to
having something more like a separate native function to check each
type.

This commit does something almost as good performance-wise as having
a native per type and typeset.  It introduces TYPECHECKER, which is a
function generator that makes a different type of function.  The "body"
is actually the type or typeset to check, and it uses a different
dispatcher for types and typesets...so even that doesn't have to be
branched on.

HAS-TYPE? is replaced by the more generalized MAYBE?, which is richer
since it is a dialect and can run test functions.