Fix bug in SCAN-NET-HEADER
A somewhat narrow bit of functionality was written as C and mixed into
the implementation of CONSTRUCT, to parse out a network header string
as a block, which would then be turned into an object.  This string
to block conversion was moved out into a routine called SCAN-NET-HEADER

The routine had to be updated for relative values and specific binding,
but it was a bit hard to follow.  That update was buggy, and a new
assert caught a problem with it.  This reorganizes the routine a bit
in order to make it more obvious.

This function should be implemented ultimately in usermode via PARSE.