Add initial support for user natives
A user native is a NATIVE function that implemented in the user mode,
and is compiled on the fly by the embedded C compiler in the
interpreter. It tries to bring the efficiency of C and flexibility of
Rebol together. Once it's compiled, it works exactly the same as the
builtin natives. User can change their implementations without
rebuilding the interpreter, which makes it easier to just implement part
of the REBOL script in C for better performance.

The embedded C compiler is TinyC (http://bellard.org/tcc), which uses
a library libtcc1.a for some of the compiled code. Thus this library has
to be distributed as the interpter to make most use of this feature
(It's possible to compile a working user native without libtcc1.a if it
doesn't require any functions provided by the library).

The preprocessed internal header file sys-core.h will be inserted into
user source code, which makes all internal functions/macros available.
External libraries can also be used if proper 'library-path' and
'library' are specified.

Here is an example:

REBOL []
c-fib: first make-native/opt [
	"N_c_fib" [
 		"nth Fibonacci Number"
		n [integer!]
	]
] {
	REBNATIVE(c_fib) {
		PARAM(1, n);

		int i = VAL_INT64(ARG(n));
		int i0 = 0, i1 = 1;
		//MARK_CELL_WRITABLE_IF_CPP_DEBUG(D_OUT);
		if (i < 0) SET_INTEGER(D_OUT, -1);
		if (i <= 1) SET_INTEGER(D_OUT, i);
		while (-- i > 0) {
			int t = i1;
			i1 = i1 + i0;
			i0 = t;
		}
		SET_INTEGER(D_OUT, i1);
		D_OUT->header.bits |= 1 << (GENERAL_VALUE_BIT + 5);
		return R_OUT;
	}
} compose [
	runtime-path (join first split-path system/options/boot %tcc)
]

fib: func [
	n [integer!]
][
	if n < 0 [return -1]
	if n <= 1 [return n]
	i0: 0
	i1: 1
	while [n > 1] [
		t: i1
		i1: i0 + i1
		i0: t
		-- n
	]
	i1
]

print ["fib 30:" c-fib 30]
print ["fib 30:" fib 30]

n-loop: 1000
c-t: dt [
	loop n-loop [c-fib 30]
]
r-t: dt [
	loop n-loop [fib 30]
]
print ["c-t:" c-t "r-t:" r-t "improvement:" r-t / c-t]

which on my system gives an improvement of about 1200x