Switch complex module call to APPLY
Originally the idea in Ren-C was to make refinements either the word
of their refinement or a BLANK! (none).  This would allow chaining
with `some-function/:refinement`, because the refinement would be
valued as the word! `refinement`.

This idea was scrapped to use TRUE and FALSE, and resultingly code
would be written as `some-function/(if refinement ['refinement])`
However, two particular calls in the module system were still relying
on the old behavior.

This updates those calls to use the easier-to-follow new APPLY syntax.
Also included are patches to WHY?, BROWSE, and tolerance of LIT-WORD!
for the module type and name.  Although WORD! is the preferred option
going forward, it creates unnecessary friction to enforce it at this
time.