Fix search for /LAST in FIND assert on short strings
An assertion caught a case where searching in short strings was causing
a numeric overflow of an unsigned length, e.g. FIND/LAST "" #"a"
The reason is because the search window had to be adjusted to account
for the length of the pattern.  This changes the code to check to see
if the pattern being searched for exceeds the length of the part of
the string being searched, and if so gives a NOT_FOUND result.