ANY-TYPE! synonym for ANY-VALUE! except in r3-legacy FUNC
While there is no UNSET! (or VOID!) type in Ren-C, an implementation
detail of functions and contexts uses a Reb_Type value to signal the
absence of a value...hence this bit is used to carry the `<opt>`
optionality status of a parameter.  (This is an optimization, and is
becoming more obviously so in light of design for user-defined types.)

However, this trick was exploited to implement the ANY-TYPE! typeset
for r3-legacy.  This made a user facing typeset that contained a
"non type", which was causing an assertion to fire when the ANY-TYPE!
typeset from r3-legacy was shown (since there was no symbol).

This changes the trick so that the ANY-TYPE! typeset is just a synonym
for ANY-VALUE!.  The ability to allow for optional arguments to
functions is done with a different compatibility hack--the r3-legacy
FUNC and FUNCTION constructs explicitly look for `[any-type!]`, and
turn it into `[<opt> any-value!]`.

The change also eliminates the ANY-SOMETHING! and ANY-NOTHING! typesets
because of the inability for typesets to carry the idea of no type.