GC sweep optimization (and comment updates)
This moves the common header bits between REBVAL and REBSER a little,
such that the lowest 3 bits (NOT_END, CELL_MASK, and *MANAGED) could
be handled by way of a switch statement in the sweep routine.  This
reduces the number of tests that need to be done by just branching on
all three bits at once.

(This could be extended to 16 cases instead of 8, by also including
the mark/unmarked bit, which may be worth it in the future.  The bit
is in the right place for it if so.)

While taking care of that, this also includes a batch of updates to
out of date comments.