Introduce "managed handles" with a GC cleanup function
This creates a new kind of "Managed" HANDLE!, which is distinct from
the traditional "Simple" HANDLE! (that was just an opaque pointer to
some data the system knew how to interpret, but that the user was
not supposed to see).

A managed handle uses a singular array--which is the size of a REBSER
node--in order to engage the garbage collector.  When the handle is
GC'd, it calls a hook function that receives the handle's REBVAL*
to do a finalization.

As an added advantage to managed handles, changing their content will
change all the references to that handle.  This means a pointer can be
moved or reallocated safely, even if several copies of that handle
REBVAL have been made.