No literal blocks in condition of IF/UNLESS/EITHER/CASE
A very common mistake for new users is to place a BLOCK! around a
conditional expression.  Since blocks are always TRUE, this leads to
confusion and bugs.

(It's an especially easy mistake to make when converting a WHILE into
an IF.)

Mechanically, functions in Ren-C are able to distinguish when a value
received as an argument was the product of an evaluation vs. literal
at the callsite.  So it can tell the difference between:

    x: [something]
    if x [print "legal case, evaluated"]

    if [something] [print "illegal case, literal"]

This distinction is used to generate an error.  Bypassing the check
can be done with a GROUP!:

    if ([something]) [print "now the block is an evaluation product"]

This extends the check to CASE, which is likely to help catch casual
errors where a body block is accidentally serving as a condition.

Also included is a check on SWITCH to not accept literal blocks as
their argument, as this is likely not what was intended (and perhaps
just a monkey-see monkey-do repeat of a block if seen on WHILE, etc.)

Reorders some errors just to avoid hitting the script error limit;
more generalized workaround pending from @ShixinZeng.