Fix build with C++, including on OSX
bool and REBOOL are not compatible.

Because <stdlib.h> declares qsort_r has a linkage of extern "C",
unresolved references are found in other modules (e.g. _T_Array). This
problem wasn't seen on Linux because the prototype of qsort is like:

       void qsort(void *base, size_t nmemb, size_t size,
                  int (*compar)(const void *, const void *));

       void qsort_r(void *base, size_t nmemb, size_t size,
                  int (*compar)(const void *, const void *, void *),
                  void *arg);

which is different from the one defined in this file, which actually
declares an overloaded version of qsort, and the one defined here still
has the C++ linkage.

However, on OSX, the prototype declared in <stdlib.h> matches the one
here exactly, and thus makes it have the C linkage.