Support cross-compiling for windows using cmake
When building with TCC, it needs to be a cross-compiler, and
present as ${CMAKE_CURRENT_BINARY_DIR}/cross-tcc

To compile for 64-bit windows:
$mkdir build
$cd build
$cmake ../make -DR3_OS_ID=0.3.40 -DCMAKE_TOOLCHAIN_FILE=../make/Toolchain-cross-mingw64-linux.cmake

Note that to build 32-bit version binaries correctly, changing
the OS_ID is not enough.  CFLAGS, CXXFLAGS, and the
CMAKE_ASM_FLAGS must be set to -m32.

Also sets ASM_FLAGS for 32-bit building on Mac