Split FILE! and URL! path dispatch, bugfix
A usage of the Val_Init_Series macro tried to write to pvs->store as
an output parameter while reading from VAL_TYPE(pvs->value) in the same
arguments, which is a problem if pvs->store and pvs->value point to
the same place.

This corrects the issue as well as clarifies the code slightly regarding
why the VAL_TYPE was needed in the first place, by tightening the
assertions and making separate entry points for PD_Url and PD_File.