User-Controllable Per-Argument Deferment of "infix" Processing
This reverses the effects of 2b4de1eca0bdb81e1c2c40c16d0cd7ea9f1b4024.

Instead of having a special rule applied to how "enfixed" functions
gather their right hand arguments, any function may now <defer> any
one of its arguments w.r.t. lookahead processing.  For example:

    >> foo: func [x [integer!]] [print x | return x * 10]

    >> foo 1 + 2
    3
    == 30

    >> bar: func [x [<defer> integer!]] [print x | return x * 10]

    >> bar 1 + 2
    1
    == 12

When this deferment of lookahead is applied to the right-hand side of
infix operations, it gives the appearance of "left to right" evaluation
which was seen in R3-Alpha.

However, since Ren-C opened up the ability to gather a left-hand
"lookback" argument to any function (of any arity), this generalization
allows all operations to choose whether they actually want this
behavior.  Some do not, such as `x: default 1 + 2`... where DEFAULT
is an infix operation that quotes its left hand argument (a SET-WORD!
or SET-PATH!) yet wishes to normally evaluate the second argument,
as opposed to be interpreted as `(x: default 1) + 2`.

Looking into this feature and what it might mean to defer the left
hand argument of an enfixed function, tricky code was written to
give meaning to the deferment of the left hand side of an enfixed
function.  This meant it would try its best to complete an expression
on the left before applying the operator:

    >> baz: enfix func [arg1 [<defer> integer!] arg2 [integer!]] [
           arg1 * arg2
       ]

    >> add 1 2 baz 3
    == 9 ;-- defers (2 baz 3), since (add 1 2) is complete

    >> add 1 baz 2 3
    == 5 ;-- can't defer (1 baz 2) since (add 1) is incomplete

Although this added complexity to the evaluator, it meant deleting a
broken version of the feature known as the "punctuator"...resulting in
an overall streamlining of the logic.