Cleanup: Remove dual senses of several DO_FLAGs
In the early creation of the modern `struct Reb_Frame`-based evaluator,
a DO_FLAG was made for both senses of several tests.  This way instead
of a callsite saying `Do_Core(..., DO_FLAG_NO_LOOKAHEAD)` or
`Do_Core(..., 0)` to imply the default of lookahead, a flag was made
to note lookahead at the callsite `Do_Core(..., DO_FLAG_LOOKAHEAD)`.
These flags were then tested to make sure only one was set.

At the time, this was good for documentation at the callsite.  But
since then the entry points to the evaluator have matured to where it
is generally not ambiguous.  The waste of flag bits, as well as
complicating the work done need to flip a bit, is no longer worth it.

If two things are mutually exclusive it's easier to represent them with
a single bit.  Given that it's overkill at this point in time, this
collapses the bits to name the uncommon cases...for the sake of
reducing overall code.