Eliminate implicit VARARGS! chaining
VARARGS! support three fundamental operations: TAIL?, TAKE, and FIRST.
(The FIRST operation is only available if the variadic operator is
"hard quoting".)

@giuliolunati noticed a problem with the interaction between TAIL? and
an experimental feature of implicit chaining of VARARGS!.  The
problem is that the TAIL? operation cannot perform an evaluation, yet
an evaluation is necessary in order to determine whether an expression
is going to evaluate to a VARARGS! which may need to be chained or not.
Hence TAIL? could be inconsistent with a subsequent TAKE (or calls
from TAIL? at different stack levels).  e.g.

    f: func [x [any-value! <...>]] [probe tail? x]
    g: func [y [any-value! <...>]] [probe tail? y | f y]
    g

gives:

    true
    false

This compounded an existing technical problem: namely, if VARARGS! are
a meaningful Rebol value in their own right...how would you ever pass
one to a variadic routine if it treats them as a chain?

Viewed in light of the need for clarity and stability of the feature
due to its complex interactions with other parts of the evaluator, the
chaining is a liability.  It was also somewhat twisty/brittle code,
developed when the experiment of variadics was first being tried out.

This removes the chaining feature.  A substitute for chaining which
will work for the most common case is to use APPLY.

    f: func [x [any-value! <...>]] [probe tail? x]
    g: func [y [any-value! <...>]] [probe tail? y | apply :f [x: y]
    g

gives:

    true
    true

Passing a VARARGS! argument through ordinary invocation to a variadic
function will not chain, and instead the TAKE of the variadic parameter
will return the literal VARARGS! from the callsite itself.  If clients
wish to interpret that as a chain they can do so themselves; but there
is no way to do this mechanically without compromising the TAIL? and
TAKE parity.