Preserve knowledge of if current function is lookback in call
Previously, when a function was being invoked in a "lookback" mode,
the fact that it had been invoked via lookback was forgotten after the
first argument was gathered.  This is because the transition from
"lookback" to "normal function" was used as a state indicator to say
that the remaining arguments should be gathered normally.

There has been some discussion of varying the behaviors of argument
gathering *besides* the first argument based on whether the invocation
was "enfixed" or not.  So this makes that kind of change possible by
adding another state to the f->refine argument gathering state list,
to take care of managing whether one is on the first argument of a
lookback function or not.

Includes some macros to make it clearer what the "special values"
mean when they are held in f->refine (as opposed to when f->refine
holds a pointer to the refinement argument slot being currently
processed.)