Use mold buffer for TCC combined source
The planned change for user natives will involve pulling together
fragments of source from indvidually declared natives.  This is much
easier to manage via the mold buffer, as opposed to trying to deal
with computing the size of the source in advance.

Since the mold buffer is UTF16, this means there has to also be a
"popping" to a UTF8 series.  This is not that big a deal, but it also
should ultimately not be necessary, as the goal is to make all string
handling internally UTF8 in the future.

Also renames the "TCC_state" variable to just "state", as capitalization
in variable names is not standard...and tcc_state might look too much
like a function from libtcc.

In addition, this uses the service routine Init_Handle_Managed()
instead of creating the singular array for the handle by hand.

Also adds the fib.r sample to the tests subdirectory.