Don't allow quoting of expression barriers
When the BAR! datatype was introduced, it was allowed to pass it to
quoted arguments.  e.g. `QUOTE |` would work.  However, as barriers
have become more widely used, it has become apparent that their utility
is greater when they really offer a guarantee that's as good as if you
had written `(QUOTE)`.  It's possible to obtain a literal expression
barrier other ways, such as with '| or FIRST [|], so this moves to
full enforcement of | as a barrier in parameter evaluation.