Update evaluator to properly handle postfix, make ? postfix
While postfix operators were possible in theory, none existed...so there
was a bug in the handling of them if they were the last value in a
series.  This does a light reorganization of the evaluator to correct
that issue and get rid of some redundant code.

Given the ability to defer the left hand argument in postfix, then the
concept becomes much more interesting--because it means a complete
expression can be gathered before it is applied.  This uses deferment
as a trial for ? to be a postfix test for ANY-VALUE?:

    >> blk: [1 2 3 4]

    >> while [x: take blk ?] [print x]
    1
    2
    3
    4

This replaces the previous experiment of ? as a prefix synonym for
ANY-VALUE?.

Also repositions the test for variadic parameters to give better errors,
and updates some comments.