Preliminary steps to eliminating RXIARG/RXIFRM
Before Rebol was open sourced, a small group of APIs were exposed for
implementing functions that could be called from Rebol called COMMAND!.
While not having access to the structural definitions of REBVALs or
REBSERs, a small set of functions known as the "RL_API" gave a limited
ability to inspect and modify them.

REBSER was exported as an "opaque type", so that clients saw `REBSER*`
as equivalent to a `void*`  The code inside a COMMAND! could hold
these pointers in stack variables indefinitely, and the routines that
provided infrastructure within the interpreter to handle things like
protection from garbage collection was not exported.

REBVAL was not exported as an opaque type, rather it was proxied into
something called an "RXIARG".  The RXIARG was a non-opaque structure
that tried to mirror a subset of potential REBVALs.  This required
copying back and forth between the RXIARG and REBVALs in a process that
was awkward and error prone.

Now that Rebol is open source, it makes more sense to divide the C API
into two categories.  One would be the internal API...in which the
full breadth of concerns and implementation tools for writing native
interpreter code is exposed.  (This internal API is now being used by
"user natives", for instance.)  It is likely that codebases like
R3/View and the implementation of wrappers like Ren-Cpp will want to
have this level of access as well.

The other category is a layer which some have requested, which would
permit easy embedding of Rebol into C codebases without having to
worry about things like PUSH_GUARD_SERIES() or RELVAL-vs-REBVAL, or
REBCTX-vs-REBARR, specifiers, end markers, etc.  This API would feel
more like using the Rebol evaluator itself from C...where the only
basic datatype to be concerned with is the "value" (or REBVAL).

This change begins a hopefully-gentle transition into coming up with
that split.  It eliminates the RXIARG structure in favor of adding
accessors to the RL_API.  Some of these accessors "break the rules"
in terms of providing intimate access to types that belong in the
internal API only, however these are transitional for convenience...
and basically all of today's code using these routines probably
really want to use the internal API anyway.

In the meantime, it offers a baby step into what it might like to work
at an API level without a zoo of datatypes mirroring the internal API.
As an example, the "RXIFRM" data type for representing a function
frame is replaced with simply a REBVAL of a FRAME!.  Eventually, all
the RL_API routines should go that route--with those needing control
at the REBSER and lower level using the internal API.