Optimization for empty user function bodies
If you write something like `foo: func [...] []`, this substitutes the
ordinary dispatcher with one that is an absolute no-op and returns
void.  This means there's no invocation equivalent to `DO []`, and
there is also no overhead for frame reification.

While this may seem like an uncommon case to optimize for, it actually
does have an application, exemplified by the default definition for
ASSERT...which is an empty body.  It's only when you are in "debug mode"
(currently the default) that ASSERT gets HIJACKed with an actual
implementation.

So in addition to helping optimize the rare case where a body ends up
being a no-op due to a COMPOSE or other random reason, this makes it
possible to create low-overhead stubs for instrumentation...which can
be hijacked with implementations when needed.

This also switches around the place where analysis is done for picking
from the optimized dispatchers for "plain functions", and also gives
the category of "plain" the new name "interpreted functions".  ("User"
function doesn't cover it, in the sense that Mezzanines would then
also be "User" functions...)