Add <with> to FUNC, eliminate <no-return> and <no-leave>
Ren-C replaced FUNCTION's /EXTERN refinement with the idea of having
`<with>` in function specs, paralleling how locals are specified.  This
adds `<with>` to FUNC, even though there is no "locals gathering" which
would affect it in the way FUNCTION needs externs.

This allows for easier transitions between FUNC and FUNCTION modes of
coding.  The `<with>` lets you put strings after the words which are
ignored, which can help serve commentary purposes on what external
variables are for.  (The strings are thrown out and not stored anywhere
accessible to HELP, merely source comments.)

    k: 10
    foo: func [
        x [integer!]
        <with> k "number of variants"
    ][
        k: x ;-- <with> was commentary, no mechanical difference
    ]

But most interestingly, it gets rid of the awkward `<no-return>` and
`<no-leave>` refinements for telling FUNC and PROC you don't want a
definitional RETURN or LEAVE.  Simply saying `<with> return` or
`<with> leave` is enough to indicate you're bringing in those words
from the outside context, and the function should not define local
meanings for them.