Get rid of superfluous "note to programmers"
Several files had a boilerplate note about the importance of considering
multiple platforms and testing.  This isn't really a substitute for a
coding practices document and just makes the files longer.

Also eliminates the prefix form of ? (which was left in despite being
overwritten by the postfix-left-arg-deferred form in %base-infix.r)