Allow FOR-EACH to enumerate all the FUNCTION!s
This permits you to use `for-each f function! [...]` to get an instance
of each function value that was in existence at the start of the
FOR-EACH.  If in the body any functions are created, they will not
be included in the enumeration.

This means that the routine has to snapshot all the values at the
start of the routine, and protect them from GC during it.  So it is
not a particularly efficient operation.  However, it is intended
primarily for instrumentation purposes (getting tables of functions to
print out for statistics).

To faciliate this, a REBSER identifying a function's paramlist is now
identified by an ARRAY_FLAG_PARAMLIST bit.  This also addresses an
outstanding issue with using functions as a garbage collection root--
as previously, it wasn't possible to tell if a REBSER was a paramlist
or not.