Factor %n-control.c into files by functional group
The file %n-control.c historically contained a large number of natives,
that spanned beyond "control and branching" constructs.  This creates
new files for protect routines, error routines, function generators,
and non-branching evaluation.

(Over the long-term it may be questionable if splitting out the
natives in the "n-xxx.c" files is worthwhile, or if they should be
coupled with their support routines.)

Also gets rid of the %a-stubs.c empty file.