Fix latent BLANK!-instead-of-void empty map keys, formatting
Ren-C makes it legal to store any value into a map, including a BLANK!.
As with variables, the void state is used to mean "not set" (so that's
how entries are seen as not being in the map.)

This is a change from R3-Alpha (and Red) which consider blanks (nones)
to be values that indicate an absense.

There were a few cases in conversions where blanks were being used
instead of voids.  This repairs that problem, as well as code that
will create an object from a map (although its handling of non-word
keys is questionable, this feature was in R3-Alpha).

Also brings the code formatting and style in line with today's Ren-C.