LOOP-WHILE, LOOP-UNTIL, and UNTIL-2
In the once-hotly-debated question of whether it made more sense for
UNTIL to be arity-2 (to match WHILE) and then have arity-1 forms of
both WHILE and UNTIL as "LOOP-UNTIL" and "LOOP-WHILE", it was decided
to keep UNTIL as arity-1 and put the burden of reconfiguring on the
user (and then to make sure that burden was very small).

To make that burden as small as possible, this goes ahead and does the
implementation of all four variants by the "ideal" definition.  Then
during boot it simply says:

    until-2: :until
    until: :loop-until

This can be undone by a user with `until: :until-2`.  The definitions
are used in Rebmu as UT, WH, LU, LW.

Includes some comment editing in %sys-rebval.h, and moves Reb_All to
not be platform dependent (but use platform-dependent REBUPT).