Formatting of u-compress and MSVC formatting fix
MSVC auto-formatting is bad, turn it off:

http://stackoverflow.com/questions/737222/

It had been taking the `fail (Error(...))` and turning it into
`fail(Error(...))` when that defeated part of the purpose of making it
look more like a keyword than a function call, so that failure points
would stand out.  It had other bad properties too.

This corrects that (and other outstanding instances of fail that were
not spaced) as well as having some small cleanup of the code style
in %u-compress.c.