Replace <defer> with new semantics as <tight>
R3-Alpha achieved an apparent "left-to-right" via lookahead suppression.
So when evaluating:

1 + 2 * 3

Lookahead would be enabled by default, 1 would evaluate to itself
and see an OP! in that lookahead.  During evaluation of the remaining
`[2 * 3]`, lookahead would be disabled.  (This would happen during the
right-hand side evaluation of any OP!.)  Hence it would see the 2,
which would evaluate to itself, and the addition would run--thus
finishing the first infix operation.  Lookahead would be re-enabled, so
it would see the `*`, lookahead would be disabled while evaluating the
3...and so on.

(In this case there's nothing further to see, so the expression is 9.)

This form of limiting argument evaluation was seen as a poor default
behavior for Ren-C.  So enfixing operators would not have this property
and would gather their arguments without lookahead suppression--just as
a prefix operation would.  So at one time, `1 + 2 * 3` would behave
just like `add 1 2 * 3`, which would behave the same as
`add 1 multiply 2 3`.  So it would evaluate to 7.

While more uniform, this was not backward-compatible.  So the `<defer>`
annotation was introduced, which could be marked on any parameter in
order to say that parameter should not do infix lookahead during its
evaluation.  Compatibility versions of the operators used this.

Because it was emerging as necessary for certain operators, the idea of
being able to control the "eagerness" of evaluation on the left hand
side of an enfixed operation was also considered.  Hence a meaning
for the <defer> annotation on that slot was given, to be "non-eager"
and complete as much of the left hand side as possible.  So if the
left-hand side of SOME-INFIX was deferred, while the right was not,
it would be interpreted as:

    1 + 2 foo 3 + 4 => (1 + 2) foo (3 + 4)

This commit switches the behavior so this bias for complete expressions
on the left hand side is the default...as is the bias for complete
expression on the right hand side.  Interestingly, this gives rise
to a legacy-compatible left-to-right interpretation of infix...but
for a different reason.

So in `1 + 2 * 3`, the `+` reads from `[2 * 3]` with the lookahead
fully enabled.  But when the `*` is hit, its default is to ask that
the left-hand side be as complete as possible.  This pushes evaluation
back to the 1 + 2, which completes, and then runs the multiplication
to get 9.

Now the request for shortest-possible-expression uses the `<tight>`
annotation on both left and right args.